[Core] fix compile error on Logger

+++ client/NetworkModule.cpp
@@ -81,7 +81,7 @@ void		NetworkModule::connectionCommand(Command const &command)
 	GameListCommand const &cmd = static_cast<GameListCommand const &>(command);
 	Net::Packet		packet(sizeof(uint16_t) + cmd._login.length() + sizeof(uint8_t));
 
-	packet << (sizeof(uint8_t) + cmd._login.length());
+	packet << static_cast<uint16_t>((sizeof(uint8_t) + cmd._login.length()));
 	packet << static_cast<uint8_t>(TCP::CONNECTION);
 	packet << cmd._login;
 	this->_server->handleOutputPacket(packet);
@@ -92,7 +92,7 @@ void		NetworkModule::createGameCommand(Command const &command)
 	GameListCommand const &cmd = static_cast<GameListCommand const &>(command);
 	Net::Packet		packet(sizeof(uint16_t) + sizeof(uint8_t) + sizeof(uint8_t) + cmd._map.length());
 
-	packet << (sizeof(uint8_t) + cmd.name.length());
+	packet << static_cast<uint16_t>((sizeof(uint8_t) + cmd.name.length()));
 	packet << static_cast<uint8_t>(TCP::CREATE_GAME);
 	packet << cmd._map;
 	this->_server->handleOutputPacket(packet);
@@ -102,7 +102,7 @@ void		NetworkModule::listGamesCommand(Command const &cmd)
 {
 	Net::Packet		packet(sizeof(uint16_t) + sizeof(uint8_t));
 
-	packet << (sizeof(uint8_t));
+	packet << static_cast<uint16_t>((sizeof(uint8_t)));
 	packet << static_cast<uint8_t>(TCP::LIST_GAMES);
 	this->_server->handleOutputPacket(packet);
 }
@@ -112,7 +112,7 @@ void		NetworkModule::connectGameCommand(Command const &command)
 	GameCommand const &cmd = static_cast<GameCommand const &>(command);
 	Net::Packet		packet(sizeof(uint16_t) + sizeof(uint8_t) + sizeof(uint16_t));
 
-	packet << (sizeof(uint8_t) + sizeof(uint16_t));
+	packet << static_cast<uint16_t>((sizeof(uint8_t) + sizeof(uint16_t)));
 	packet << static_cast<uint8_t>(TCP::LIST_GAMES);
 	packet << (static_cast<int16_t>(cmd.idObject)); // idObject represents here idGame
 	this->_server->handleOutputPacket(packet);
@@ -123,7 +123,7 @@ void		NetworkModule::playerCommand(Command const &command)
 	GameListCommand const &cmd = static_cast<GameListCommand const &>(command);
 	Net::Packet		packet(sizeof(uint16_t) + sizeof(uint8_t) + sizeof(uint8_t) + cmd._login.length());
 
-	packet << sizeof(uint8_t) + sizeof(uint8_t) + cmd._login.length();
+	packet << static_cast<uint16_t>(sizeof(uint8_t) + sizeof(uint8_t) + cmd._login.length());
 	packet << static_cast<uint8_t>(TCP::PLAYER);
 	packet << static_cast<uint8_t>(cmd._status);
 	packet << cmd._login;
